kernel: Link kobjs into core.elf — exec.library + kernel.resource boot!
Major milestone: core.elf now links kernel_resource.o, exec_library.o,
and task_resource.o kobjs into a single 226KB kernel image that boots
on QEMU raspi4b.

Changes:
- kernel mmakefile: Build kernel_cstart.c as part of kernel.resource kobj,
  link kobjs with -larossupport -lautoinit -llibinit -lstdc.static
- kernel_cstart.c: Use extern for SysBase/BootMsg/__aarch64_arosintern
  (defined in kernel_startup.c/kernel_globals.c), place _start in
  .text.startup section
- cpu_superstate.c: New file for AArch64 cpu_SuperState() (returns SP,
  always in EL1)
- ldscript.lds: Place .text.startup first for _start entry point
- boot/elf.c: Add R_AARCH64_ADR_GOT_PAGE and LD64_GOT_LO12_NC
  relocations, return _start symbol address from loadElf()
- boot/boot.c: Use ELF _start address as kernel entry point

Boot output:
  [BOOT:ELF] _start at 0x3bc05be8
  [BOOT] Jumping to kernel @ 0x3bc05be8
  [Kernel] AROS AArch64 Kernel (Apr 11 2026)

Next: Fix tag passing so kernel can parse boot parameters.

+++ arch/aarch64-raspi/boot/include/elf.h
@@ -16,7 +16,7 @@ extern struct bss_tracker tracker[];
 
 int getElfSize(void *elf_file, uint64_t *size_rw, uint64_t *size_ro);
 void initAllocator(uintptr_t addr_ro, uintptr_t addr_rw, uintptr_t virtoffset);
-int loadElf(void *elf_file);
+uintptr_t loadElf(void *elf_file);
 
 #define ELF64_R_SYM(val)   ((val) >> 32)
 #define ELF64_R_TYPE(val)   ((val) & 0xffffffffUL)
